+++ src/gui/gui.mjs
@@ -63,7 +63,6 @@ function getImportOpts(manifest) {
   if (manifest.catalog) {
     opts.catalog = manifest.catalog;
   }
-  opts.display_all = vars['display-all'] || vars.a || !!manifest.display_all;
   opts.quick_view = vars['quick-view'] || vars.q || !!manifest.quick_view;
   opts.target = vars.target || manifest.target || null;
   opts.name = vars.name || manifest.name || null;